refactor: some lints, TODOs, styling and file-splitting
