+++ apps/templates/debt/_list.html
@@ -10,17 +10,17 @@
         </small>
 
         <ul class="list-group mt-3">
-            <li class="list-group-item">
+            <li class="list-group-item"
+                hx-trigger="click"
+                hx-get="{% url 'htmx-money-spent-on-room' slug=room.slug %}"
+                hx-target="#debt-tab-content"
+                hx-swap="innerHTML">
                 <div class="row align-items-center">
                     <div class="col-11">
                         <small>See how much each person has paid so far</small>
                     </div>
 
-                    <div class="col-1"
-                         hx-trigger="click"
-                         hx-get="{% url 'htmx-money-spent-on-room' slug=room.slug %}"
-                         hx-target="#debt-tab-content"
-                         hx-swap="innerHTML">
+                    <div class="col-1">
                         <div class="float-end">
                             <i class="bi bi-caret-right"></i>
                         </div>
