+++ apps/transaction/views/htmx_views/transaction_detail_view.py
@@ -1,4 +1,3 @@
-from ambient_toolbox.view_layer import htmx_mixins
 from django.utils.functional import cached_property
 from django.views import generic
 from django_context_decorator import context
@@ -6,7 +5,7 @@ from django_context_decorator import context
 from apps.transaction.models import ParentTransaction
 
 
-class TransactionDetailHTMXView(htmx_mixins.HtmxResponseMixin, generic.DetailView):
+class TransactionDetailHTMXView(generic.DetailView):
     model = ParentTransaction
     context_object_name = "parent_transaction"
     template_name = "transaction/_detail.html"
@@ -20,6 +19,3 @@ class TransactionDetailHTMXView(htmx_mixins.HtmxResponseMixin, generic.DetailVie
     @cached_property
     def room(self):
         return self.object.room
-
-    def get_object(self, queryset=None):
-        return super().get_object(queryset)
